mod: the output in training set is changed from New Block Temp to New Block Rate
The each row in the training set is collected every 0.2 second. If we train the machine learning model to predict new block temp, the prediction result is only usable if the update frequency of TBC controller is also 0.2s. If we train the machine learning model to predict new block rate, then the update frequency of TBC controller can be any value.